Project Development Phase/Sprint 2/app/requirements.txt
flask
ibm_db
ibm-cos-sdk
pyjwt
requests

sib-api-v3-sdk